requirements.txt
torch
torchvision
grpcio
grpcio-tools
pillow
opencv-python-headless
kafka-python